Split errors into different files
